Auth and Users

+++ app/routers/listing.py
@@ -2,13 +2,18 @@ from fastapi import Depends, APIRouter, HTTPException, Response
 from pyodbc import Cursor
 
 from app.database import get_db
-from app.schemas import ListingCreate, Listing, ListingUpdate
+from app.schemas import ListingCreate, Listing, ListingUpdate, UserIdentifier
+from app.jwt import get_current_user
 
 router = APIRouter(prefix="/listings", tags=["Listings"])
 
 
 @router.post("/", response_model=Listing)
-async def create_listing(listing_data: ListingCreate, db: Cursor = Depends(get_db)):
+async def create_listing(
+    listing_data: ListingCreate,
+    db: Cursor = Depends(get_db),
+    current_user: UserIdentifier = Depends(get_current_user),
+):
     # Insert new listing.
     insert_query = """
         INSERT INTO listings (title, description, price, seller_id)
@@ -19,7 +24,7 @@ async def create_listing(listing_data: ListingCreate, db: Cursor = Depends(get_d
         listing_data.title,
         listing_data.description,
         listing_data.price,
-        listing_data.seller_id,
+        current_user.user_id,
     )
 
     db.commit()
@@ -35,12 +40,17 @@ async def create_listing(listing_data: ListingCreate, db: Cursor = Depends(get_d
     # Preparing output object.
     listing = vars(listing_data)
     listing["listing_id"] = listing_id
+    listing["seller_id"] = current_user.user_id
 
     return listing
 
 
 @router.get("/{listing_id}", response_model=Listing)
-async def get_listing(listing_id: int, db: Cursor = Depends(get_db)):
+async def get_listing(
+    listing_id: int,
+    db: Cursor = Depends(get_db),
+    current_user: UserIdentifier = Depends(get_current_user),
+):
     l = db.execute("SELECT * FROM listings WHERE listing_id = ?;", listing_id).fetchone()
     if not l:
         raise HTTPException(status_code=404, detail="Listing not found")
@@ -54,7 +64,11 @@ async def get_listing(listing_id: int, db: Cursor = Depends(get_db)):
 
 
 @router.get("/", response_model=list[Listing])
-async def get_listings(keyword: str | None = None, db: Cursor = Depends(get_db)):
+async def get_listings(
+    keyword: str | None = None,
+    db: Cursor = Depends(get_db),
+    current_user: UserIdentifier = Depends(get_current_user),
+):
     if keyword:
         # Use a parameterized query to avoid SQL injection
         query = """
@@ -68,8 +82,8 @@ async def get_listings(keyword: str | None = None, db: Cursor = Depends(get_db))
         db.execute(query, (keyword_pattern, keyword_pattern))
     else:
         query = "SELECT * FROM listings ORDER BY listing_id;"
+        db.execute(query)
 
-    db.execute(query)
     cols = [column[0] for column in db.description]
     listings = db.fetchall()
 
@@ -81,10 +95,15 @@ async def update_listing(
     listing_id: int,
     listing_data: ListingUpdate,
     db: Cursor = Depends(get_db),
+    current_user: UserIdentifier = Depends(get_current_user),
 ):
-    existing_listing = db.execute("SELECT * FROM listings WHERE listing_id = ?;", listing_id).fetchone()
+    existing_listing = db.execute(
+        "SELECT * FROM listings WHERE listing_id = ?;", listing_id
+    ).fetchone()
     if not existing_listing:
         raise HTTPException(status_code=404, detail="Listing not found")
+    if current_user.user_id != existing_listing.seller_id:
+        raise HTTPException(status_code=403, detail="You cannot modify other people's listings!")
 
     # Update the listing data
     update_query = """
@@ -116,11 +135,17 @@ async def update_listing(
 async def delete_listing(
     listing_id: int,
     db: Cursor = Depends(get_db),
+    current_user: UserIdentifier = Depends(get_current_user),
 ):
     # Check if the listing exists
-    existing_listing = db.execute("SELECT * FROM listings WHERE listing_id = ?;", listing_id).fetchone()
+    existing_listing = db.execute(
+        "SELECT * FROM listings WHERE listing_id = ?;", listing_id
+    ).fetchone()
     if not existing_listing:
         raise HTTPException(status_code=404, detail="Listing not found")
+    # Check if the current user can delete the listing.from
+    if current_user.user_id != existing_listing.seller_id:
+        raise HTTPException(status_code=403, detail="You cannot delete other people's listings!")
 
     # Delete the listing
     db.execute("DELETE FROM listings WHERE listing_id = ?;", listing_id)
